# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/game-sessions": {
      "post": {
        "operationId": "MultiplayerController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameSessionConnectionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new game session, with the caller as host",
        "tags": [
          "game-sessions"
        ]
      },
      "get": {
        "operationId": "MultiplayerController_list",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameSessionListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List game sessions for a project, from the caller's perspective",
        "tags": [
          "game-sessions"
        ]
      }
    },
    "/game-sessions/join-by-code": {
      "post": {
        "operationId": "MultiplayerController_joinByCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinByCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameSessionConnectionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Join an invite-code game session by its code",
        "tags": [
          "game-sessions"
        ]
      }
    },
    "/game-sessions/{sessionId}": {
      "get": {
        "operationId": "MultiplayerController_get",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameSessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch a single game session",
        "tags": [
          "game-sessions"
        ]
      },
      "patch": {
        "operationId": "MultiplayerController_update",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update game session settings (host only)",
        "tags": [
          "game-sessions"
        ]
      },
      "delete": {
        "operationId": "MultiplayerController_remove",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Close/delete a game session (host only)",
        "tags": [
          "game-sessions"
        ]
      }
    },
    "/game-sessions/{sessionId}/players": {
      "get": {
        "operationId": "MultiplayerController_players",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRosterResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Who is in a game session",
        "tags": [
          "game-sessions"
        ]
      }
    },
    "/game-sessions/{sessionId}/invite": {
      "post": {
        "operationId": "MultiplayerController_invite",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteToSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Invite someone to a session (host only)",
        "tags": [
          "game-sessions"
        ]
      }
    },
    "/game-sessions/{sessionId}/join": {
      "post": {
        "operationId": "MultiplayerController_join",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinGameSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameSessionConnectionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Join a game session as a player",
        "tags": [
          "game-sessions"
        ]
      }
    },
    "/game-sessions/{sessionId}/leave": {
      "post": {
        "operationId": "MultiplayerController_leave",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Leave a game session as a player",
        "tags": [
          "game-sessions"
        ]
      }
    },
    "/game-sessions/{sessionId}/ticket": {
      "post": {
        "operationId": "MultiplayerController_refreshTicket",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameSessionConnectionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mint a fresh connection ticket for the caller's session",
        "tags": [
          "game-sessions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Naucto API",
    "description": "The Naucto API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT : Bearer <token>"
      }
    },
    "schemas": {
...
      "CreateGameSessionDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "number",
            "description": "ID of the project this session is played on"
          },
          "title": {
            "type": "string",
            "description": "Human-readable title of the session"
          },
          "maxPlayers": {
            "type": "number",
            "description": "Maximum number of players, host included",
            "minimum": 2,
            "maximum": 16
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "FRIENDS_ONLY",
              "INVITE_CODE"
            ]
          }
        },
        "required": [
          "projectId",
          "title",
          "maxPlayers",
          "visibility"
        ]
      },
      "WebRTCOfferPeerICEServerConfig": {
        "type": "object",
        "properties": {
          "urls": {
            "type": "string"
          },
          "username": {
            "type": "object"
          },
          "credential": {
            "type": "object"
          }
        },
        "required": [
          "urls"
        ]
      },
      "WebRTCOfferPeerOptsConfig": {
        "type": "object",
        "properties": {
          "iceServers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebRTCOfferPeerICEServerConfig"
            }
          }
        },
        "required": [
          "iceServers"
        ]
      },
      "WebRTCOfferPeerOpts": {
        "type": "object",
        "properties": {
          "config": {
            "$ref": "#/components/schemas/WebRTCOfferPeerOptsConfig"
          }
        },
        "required": [
          "config"
        ]
      },
      "WebRTCOfferDto": {
        "type": "object",
        "properties": {
          "signaling": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxConns": {
            "type": "number"
          },
          "peerOpts": {
            "$ref": "#/components/schemas/WebRTCOfferPeerOpts"
          }
        },
        "required": [
          "signaling",
          "maxConns",
          "peerOpts"
        ]
      },
      "GameSessionConnectionResponseDto": {
        "type": "object",
        "properties": {
          "sessionUuid": {
            "type": "string"
          },
          "playerId": {
            "type": "number",
            "description": "Player id assigned to this connection"
          },
          "webrtcConfig": {
            "$ref": "#/components/schemas/WebRTCOfferDto"
          },
          "connectionTicket": {
            "type": "string",
            "description": "Short-lived signed ticket presented on the WebRTC connection"
          },
          "joinCode": {
            "type": "string",
            "description": "Join code to share, present only when hosting an INVITE_CODE session"
          }
        },
        "required": [
          "sessionUuid",
          "playerId",
          "webrtcConfig",
          "connectionTicket"
        ]
      },
      "JoinByCodeDto": {
        "type": "object",
        "properties": {
          "joinCode": {
            "type": "string",
            "description": "Invite code of the session to join"
          },
          "editorTest": {
            "type": "boolean",
            "description": "Set by the game editor to allow a self-join for solo testing"
          }
        },
        "required": [
          "joinCode"
        ]
      },
      "GameSessionResponseDto": {
        "type": "object",
        "properties": {
          "sessionUuid": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "FRIENDS_ONLY",
              "INVITE_CODE"
            ]
          },
          "hostId": {
            "type": "number",
            "description": "ID of the host user"
          },
          "hostUsername": {
            "type": "string",
            "description": "Username of the host"
          },
          "hostNickname": {
            "type": "string",
            "description": "Display nickname of the host, if set"
          },
          "projectName": {
            "type": "string",
            "description": "Name of the game (project) being played"
          },
          "maxPlayers": {
            "type": "number",
            "description": "Maximum number of players, host included"
          },
          "playerCount": {
            "type": "number",
            "description": "Current number of players, host included"
          }
        },
        "required": [
          "sessionUuid",
          "title",
          "visibility",
          "hostId",
          "hostUsername",
          "projectName",
          "maxPlayers",
          "playerCount"
        ]
      },
      "GameSessionListResponseDto": {
        "type": "object",
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameSessionResponseDto"
            }
          }
        },
        "required": [
          "sessions"
        ]
      },
      "SessionPlayerDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "User ID",
            "example": 42
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "alice"
          },
          "nickname": {
            "type": "string",
            "description": "Display nickname, if set",
            "nullable": true
          },
          "host": {
            "type": "boolean",
            "description": "Whether this player is hosting",
            "example": false
          }
        },
        "required": [
          "userId",
          "username",
          "host"
        ]
      },
      "SessionRosterResponseDto": {
        "type": "object",
        "properties": {
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionPlayerDto"
            }
          },
          "maxPlayers": {
            "type": "number",
            "description": "Slots the session was opened with",
            "example": 4
          }
        },
        "required": [
          "players",
          "maxPlayers"
        ]
      },
      "InviteToSessionDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "Who to invite",
            "example": 42
          }
        },
        "required": [
          "userId"
        ]
      },
      "UpdateGameSessionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "New title of the session"
          },
          "maxPlayers": {
            "type": "number",
            "description": "New maximum number of players, host included",
            "minimum": 2,
            "maximum": 16
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "FRIENDS_ONLY",
              "INVITE_CODE"
            ]
          }
        }
      },
      "JoinGameSessionDto": {
        "type": "object",
        "properties": {
          "joinCode": {
            "type": "string",
            "description": "Join code, required for INVITE_CODE sessions"
          },
          "editorTest": {
            "type": "boolean",
            "description": "Set by the game editor to allow a self-join for solo testing"
          }
        }
      },
...
    }
  }
}